changed amount data types to decimal

+++ db/migrate/02_create_budgets.rb
@@ -2,7 +2,7 @@ class CreateBudgets < ActiveRecord::Migration[5.2]
   def change
     create_table :budgets do |t|
       t.string :month
-      t.integer :amount
+      t.numeric :amount
       t.integer :user_id
       t.timestamps
     end
