+++ db/migrate/04_create_expenses.rb
@@ -2,7 +2,7 @@ class CreateExpenses < ActiveRecord::Migration[5.2]
   def change
     create_table :expenses do |t|
       t.string :name
-      t.integer :amount
+      t.numeric :amount
       t.integer :budget_id
       t.string :category
       t.timestamps
